much stronger

+++ src/variables.hpp
@@ -48,8 +48,8 @@ constexpr uint64_t moveorder_bit[64] = {
 };
 
 
-ankerl::unordered_dense::map<std::pair<uint64_t, uint64_t>, std::pair<int, int>> transpose_table;
-ankerl::unordered_dense::map<std::pair<uint64_t, uint64_t>, std::pair<int, int>> former_transpose_table;
+ankerl::unordered_dense::map<std::pair<uint64_t, uint64_t>, std::pair<int64_t, int64_t>> transpose_table;
+ankerl::unordered_dense::map<std::pair<uint64_t, uint64_t>, std::pair<int64_t, int64_t>> former_transpose_table;
 
 ankerl::unordered_dense::map<std::pair<uint64_t, uint64_t>, int> fixedstone_table;
